Fix MJML rendering on docs branch (#503)
On the docs branch, we weren't caching the `node_modules` directory as part of
the build, which exposed the fact that the build was using the wrong Gradle
task to set up the MJML renderer. Use the correct task, and also start caching
`node_modules` so docs get rendered more quickly.