SW-1291 Support multiple Balena fleet IDs
Our internal device managers are running in a different Balena fleet than the
customer-owned ones, but we want both fleets to report sensor data to the prod
terraware-server instance. Allow multiple fleet IDs to be configured.

We'll update the server environments to include both the old-style and new-style
values before deploying this change, then remove the old-style values once this
is rolled out.