SW-1310 Regenerate device manager credentials (#514)
Upgrading Keycloak changed the format of authentication URLs, and since the URLs
are embedded in offline refresh tokens, existing tokens won't work.

Add a button to the device managers page in the admin UI to generate fresh tokens
for all device managers that are associated with users. This is only available for
super-admins since it cuts across all organizations.

Also add a button to generate a new token for a single device manager.