SW-1322 Add viability test selection
Add a `selected` field to the viability test objects in the accession API
payloads. Selecting a viability test doesn't actually do anything yet; this is
just the implementation of the ability to select and deselect tests. Subsequent
changes will add logic to use the results of the selected test as the accession's
overall viability.

At most one test may be marked as selected for a particular accession. It is
completely valid for none of an accession's tests to be selected.