SW-1322 Add viability test selection as a search field
Add a boolean search field that acts the same as the `selected` field on viability
test API payloads.

Under the covers, this is a new kind of search field that wraps an `EXISTS()` SQL
expression, since field selection is actually stored in a separate table rather
than as a simple boolean column on `viability_tests`. From the client's point of
view, though, it acts like any other boolean search field.