SW-1341 Add accession counts by state to summary API (#529)
To allow the web app to show a breakdown of active accessions by current state,
add a field to the `/api/v1/seedbank/summary` response that includes the count
for each state that's considered active. All active states are always included
whether or not there are any accessions in that state.