SW-1350 Migrate to accession_collectors table 1/4 (#534)
We're getting rid of the distinction between primary and secondary collectors,
replacing them with a simple ordered list of collectors for each accession.

We'll migrate collector names to a single unified table. The migration will
happen in four steps:

1. Introduce the new table and start writing to it. Continue writing the old
   column/table as well, and continue reading from the old column/table.
   (This PR.)
2. Copy existing data to the new table. Start reading the new table.
3. Stop writing the old column/table.
4. Drop the old column/table.

This sequence eliminates any time windows when requests come in during the
middle of code deployment and we end up with data inconsistencies.

The API will be updated as well, but that'll happen after the data migration.