SW-1350 Migrate to accession_collectors table 2/4 (#535)
This is step 2 of a sequence of changes to move the collector names from
`accessions.primary_collector_name` and `accession_secondary_collectors` to a
new `accession_collectors` table.

In this step, we backfill `accession_collectors` from the existing data and
update the code to read from it instead of from the old locations. We continue
to write both locations so there isn't data inconsistency while this is in the
middle of being deployed and there are still requests being handled by the
previous code version.

Step 3 will stop writing to the old column and table, and step 4 will drop them.