SW-1350 Migrate to accession_collectors table 3/4 (#539)
`accessions.primary_collector_name` and `accession_secondary_collectors` are no
longer queried anywhere, so stop writing to them.

The fourth and final step in this migration series will drop them, but this step
needs to be deployed first so that dropping them doesn't cause failures while the
code is being deployed and the cluster is running a mix of old and new releases.