SW-1228 Add endpoint to fetch accession history
In the new version of the web app, we need to show a tab with the history of
certain kinds of changes to the accession. Add an endpoint that returns a
time-ordered list of items to show in that UI.

This initial implementation only includes accession creation and state change
events, but those should be enough to start building the UI. Additional types
of history will be added later, but the same API endpoint will return them.