SW-1398 Track user ID in accession state history 1/2
We already keep track of the history of accession state changes; this was needed
to support the "changes since last week" summary statistics. Those statistics are
gone in the new version of the web app, but we're going to need to start showing
the history of changes to an accession, including which user made each change.

Add a user ID column to `accession_state_history` and start populating it.

The second change in this series will backfill the existing history entries and
make the column non-nullable. These changes need to be deployed in sequence to
maintain bidirectional compatibility during code deployment.