Enable cascading deletes on accession child tables (#548)
The initial version of the app didn't allow deleting accessions, so I erred on the
side of safety and didn't put `ON DELETE CASCADE` clauses on the foreign-key
constraints on the various child tables.

Now we support deleting accessions, though, so it makes sense to modify the
constraints to auto-delete rows in tables that exist purely to hold lists of
values in accessions. There is never a situation where, e.g., we'll want the
existence of a viability test on an accession to cause us to be unable to delete
the accession.

However, we _do_ want that behavior for accession photos: if we haven't deleted
an accession's photos from the file store yet, we shouldn't be able to delete the
accession since it would cause the files to be orphaned.