SW-1406 Add collection source field to API
The upcoming version of the web app is replacing the "source plant origin"
accession field with "collection source." Add the new field.

The list of allowed values is different; it's possible to losslessly map the
old values to the new ones but not the other way around. So we can't do this as
a purely server-side data model change that supports both fields interchangeably.
Once clients have been updated to use the new field, existing accessions that have
source plant origin values will be backfilled and the old field will be removed.