SW-1228 Add withdrawals to accession history (#550)
Include withdrawal events in the `/api/v1/seedbank/accessions/{id}/history` API.

Currently these events are attributed to the person whose name is entered into the
"staff responsible" field in the app, which may or may not be a Terraware user's
name. In the near future we'll be constraining that field to the list of
organization members, but with the current data model, we have to treat it as an
optional text value.