SW-1390 More detailed collection site data 1/3 (#551)
The upcoming version of the app has a new, more detailed, set of fields to
describe the locations where seeds were collected. Update the API and data model
to reflect the change.

Completely new fields in the accession payloads:

* `collectionSiteCity`
* `collectionSiteCountryCode` (must be a valid two-letter country code, same as
  country code field on organization payloads)
* `collectionSiteCountrySubdivision` (freeform text, not restricted to the list
  of states/provinces in the `country_subdivisions` table)

New names for existing fields (both names are populated in GET payloads, and if
both are preent in POST/PUT payloads, the old names take precedence):

* `collectionSiteLandowner` (previously `landowner`)
* `collectionSiteName` (previously `siteLocation`)
* `collectionSiteNotes` (previously `environmentalNotes`)

Similarly, in the search API, the existing field names continue to work but there
are new ones:

* `collectionSiteCity`
* `collectionSiteCountry` (a sublist pointing to the countries search table, so
  you can do, e.g., `collectionSiteCountry_name` to get the English name of the
  country)
* `collectionSiteCountryCode`
* `collectionSiteCountrySubdivision`
* `collectionSiteLandowner` (previously `landowner`)
* `collectionSiteName` (previously `siteLocation`)
* `collectionSiteNotes` (previously `collectionNotes`)

In the database, we need to rename `accessions.environmental_notes`; the other
two values that are renamed in the API were already correctly named in the
database. This initial change adds `accessions.collection_site_notes` and writes
to both fields for bidirectional compatibility.