Alphabetize entries in some source files (#556)
We have a handful of source files that have lists of tables or methods that aren't
trivially short but make sense to keep grouped together. They weren't ordered in a
way that made them easy to navigate or that made it obvious exactly where to add
new entries. Update them to sort the relevant parts alphabetically.

No functional changes here, just shuffling of names.