SW-1388 Refactor SearchService to expose subquery
To calculate accession stats based on search filters, we will need to construct a
"get the summary stats" query that includes the underlying search query. Rework
the query construction in `SearchService` so that it exposes a method that returns
the `NestedQueryBuilder`.

No change in behavior here (tests continue to pass as-is).