SW-1388 Query summary statistics all at once (#558)
We will need to generate some of the same summary statistics for search results
that we do on the accessions dashboard page. Rework the statistics computation
so we'll be able to pass it a query generated by `SearchService`.

As part of this change, we now calculate all the statistics in a single query
rather than using a separate query for each one. This makes the query a little
ugly but will improve performance when we're calculating the stats over a complex
user-defined query.