SW-1426 Add new accession states
The new version of the app is going to switch to a simplified list of accession
states. Some of them are the same as existing states, but there are also a few
brand-new ones. Add them to the list of states.

No accessions should ever be in these states at the moment, but with this change,
they'll be included in the OpenAPI schema and will also be returned by API
endpoints that list all the values of enums.