SW-1425 Add manual state flag to accessions
The new version of the app will allow users to manually edit accession states.
There will still be some automatic state transitions but far fewer than before,
and they will follow a different set of rules.

Add support for manual state management in the data model, including the rules
for automatic transitions. This is tracked on a per-accession basis and an
accession can be moved back and forth between manual and automatic state
management, though the conversion is a bit lossy since the list of valid states
isn't the same in the two worlds.

This only changes the data model and the transition rules; the API behavior
remains the same as before. If you `GET` an accession whose state was manually
set to one of the newly-introduced states, it will be automatically converted
to one of the previously-existing states for backward compatibility.