Replace device info with accession data source 1/2
Previously, we captured a bunch of diagnostic information about the mobile
device when accessions were created by the seed collector app. We have never
used that information for anything other than inferring whether or not an
accession came from the mobile app. We already expose that in the API as the
`source` field on the accession payload; the value of that field is inferred
from the presence or absence of mobile device details.

We do still want to keep track of which accessions were created via the seed
collector app, and we'll also want to know which ones were created using CSV
upload when we add that feature.

Update the data model to explicitly store the `source` field rather than
inferring it, and add a third possible value to represent importing from CSV
or other kinds of files.

This removes the `deviceInfo` payload field and adds the `source` field to
the accession creation payload. If not present, it defaults to `Web`, mirroring
the current behavior when no device info is present.

This first change removes all the references to the `app_devices` table but
doesn't drop the table so as to not break the previous code version during
deployment; a followup change will backfill the data sources of existing
accessions and drop `app_devices`.