Fix bug in sort order generation in search queries
The search code generates `ORDER BY` clauses that use field positions to refer to
fields that are also in the `SELECT` clause, e.g., 'SELECT a, b FROM c ORDER BY 1`
to sort by thee `b` column. It thus keeps track of what order the fields are added
to the `SELECT` clause so it can use the correct position numbers.

Unfortunately, the code that was actually generating the `SELECT` clause would
always put all the scalar fields first in the list, before all the sublist fields,
even if the caller had specified that the sublist fields should come first. Thus
the `ORDER BY` clause would sometimes use the wrong positions, causing the results
to be incorrectly sorted.

Modify the code to put the fields in the positions the `ORDER BY` construction
code expects them to be, and add a test case that verifies the fix.