SW-1425 Introduce v2 accession API endpoints (#575)
To support the different behaviors of accessions in the upcoming version of the
seed management web app, introduce an alternate set of API endpoints. The new
endpoints will behave according to the specifications for the new version, even
for existing accessions. They omit fields that are being removed, and rename some
fields whose names no longer reflect user-facing terminology.

This initial change just covers a portion of the top-level accession payloads and
provides the ability for the client to explicitly change the state of an
accession. It does not change any of the child objects such as viability tests or
withdrawals, and it doesn't change the top-level accession fields related to
how we track changes to seed quantities or how we track overall viability levels.
Those changes will come later.

See https://github.com/terraware/terraware-server/pull/575 for a detailed list of
payload differences from v1; it's omitted here because it's pretty long.