SW-1357 Add latest observed quantity/time to v2 API
In order to display a warning to the user when a change in the date of a
withdrawal would affect the remaining quantity of an accession, the frontend
code needs to know the time of the most recent user-observed accession
quantity. That timestamp won't be shown to users, just used by the frontend code.

Add `latestObservedQuantity` and `latestObservedTime` fields to the v2
accession response payloads. These are computed values, so they're not included
in the request payloads.

The v2 API doesn't support directly updating the remaining quantity yet, so for
now, these two new fields are always calculated from the v1 quantity, similarly
to the way we do an on-the-fly mapping of v1 accession states to v2 ones.
Subsequent changes will start to introduce the write path for v2-style
quantities.