SW-1537 Define v2 withdrawals endpoints (#577)
We want the v2 accessions API to be more fine-grained than the v1 one. Start
moving in that direction by introducing a separate set of endpoints for reading
and writing withdrawals.

`PUT /api/v2/accessions/{id}` will no longer create, delete, or update
withdrawals.

**The POST and PUT endpoints don't work yet!** They can't be implemented until
the broader updates to quantity tracking are in place. This change just defines
the URL paths and payload formats so client developers can see what they look
like.

The GET and DELETE endpoints don't depend on the new quantity tracking logic, so
they're implemented here, though the quantity values are copied verbatim from
the v1 data and may not be consistent with v2 semantics.

The payloads defined here differ from the ones that are included as children of
the accession object in the v1 API. There are now separate payloads for GET,
POST, and PUT operations, and some fields have been added and removed.

See https://github.com/terraware/terraware-server/pull/577 for the list of field
differences.