SW-1537 Calculate observed quantity in v1->v2 conversion (#587)
If you read an accession with the v2 API and it was written to the database with
the v1 API, or vice versa, we do an on-the-fly conversion to the appropriate
style.

Previously, this was done in the controller and payload classes, but as the new
quantity tracking logic rolls out, the conversion is going to get complex enough
and involve enough business logic that it makes more sense to do it closer to the
rest of the business logic. Move the calculation to `AccessionModel` and add
fields to hold the observed quantity and time.