SW-1537 Initial implementation of v2 quantity tracking (#588)
This updates the v2 API so that it handles seed quantities closer to the way we're
going to be handling them going forward. The product spec has the full details,
but the big-picture differences, briefly:

* The "remaining quantity" accession field is now read/write: it's how users
  specify the initial quantity of the accession, and it gets updated as
  withdrawals are created, but users can also override it at any time.
* There is no longer a behavior difference between accessions with count-based and
  weight-based quantities, and in fact, you can switch an accession back and forth
  between the two (subject to constraints).
* Withdrawals always take the withdrawal quantity as input and no longer need the
  remaining quantity. The quantity can be in either seeds or weight, though mixing
  quantity types requires the subset weight/count to be specified.

THIS IS A WORK IN PROGRESS! Not all of the logic is in place yet, and the
conversion between v1 and v2 quantity tracking may produce incorrect results or
outright fail in some cases. But it works well enough to be able to support
initial frontend development against the v2 API.