SW-1561 Honor search scopes on all tables (#590)
`/api/v1/seedbank/values/all` has a concept of a "search scope" where we limit
results to the ones relevant to a particular organization or facility. The scope
was being honored for searches on the accessions table, but not for searches on
other tables. So if you asked for, e.g., all the values of `bags_number` for a
given facility ID, you'd actually get all the bag numbers across all the
facilities you had access to.

Eventually we're going to want to revamp how that endpoint works (SW-552) but
in the meantime, the API ought to work as advertised.