Update create_accessions.py to send v2 requests (#591)
Update the test script that creates random accessions so it can optionally do so
using the v2 API endpoints. This allows a limited form of fuzz testing of the v2
API: it doesn't check for correctness but can detect if the request handler bombs
out on what ought to be valid input.