SW-1574 Add viability percent to v2 accession API
Add a read/write field for the accession's viability percentage. This may be
edited by the user or autofilled on the client side based on test results.

v1 interoperability note: This is a calculated field (`totalViabilityPercent`)
in the v1 API.  If you do a v1 PUT, any user-edited value from a previous v2
PUT will be overwritten. A v1 GET will always return the calculated value.