SW-1537 Calculate estimated weight of accession (#606)
In the v2 web app, we need to be able to show the estimated seed count for
weight-based accessions as well as the estimated weight for count-based ones,
assuming the user has entered subset measurements.

We already calculate the estimated seed count in v1, but it's based on the initial
quantity, which is going away as a concept in v2.

Add estimated weight to the v2 accession GET payload and the search API, and
update the estimated count calculation to use the remaining quantity on v2.

For consistency with the subset count and weight values, rename
`estimatedSeedCount` to `estimatedCount` in the accession payload, and make
`estimatedCount` available as a search field; the existing search field is called
`estimatedSeedsIncoming` which will no longer be true in v2 once the estimate is
based on the remaining quantity.