SW-1274 Make plantsCollectedFrom a single value
We decided against changing this to a range in the new version of the mobile app.
However, we still want the field name to refer to "plants" rather than "trees."

The database hadn't yet been updated to store a range, so this is purely a change
in API payloads and search field names.