SW-1600 Keep track of minimum mobile app versions
To support forced updates of mobile apps if we make breaking API changes, as
well as notifying users when there significant new versions of mobile apps are
released, add an API endpoint that returns a list of our apps with a minimum
and recommended version number for each.

The list of app versions is maintained in a new repeatable database migration
so we can ensure it matches the server code as we update the API. But having to
update server-side database migrations would be annoying during mobile app
development, so the list is also editable in the admin UI for super-admin users.
Edits made in the admin UI are overwritten when the migration script is updated.