SW-1596 Substrates and treatments for v2 viability tests (#612)
In v2, we are expanding the list of viability test treatments and changing the
substrates so that there are different lists of valid values depending on test
type. Most of the existing substrate values are also being renamed, e.g., "Agar
Petri Dish" is now "Agar."

Update the underlying data model so it uses the v2 values. Continue to accept
and return the existing values from the v1 accessions endpoints.

In cases where there is no unambiguous mapping between v1 and v2 treatments,
we remove the value rather than defaulting to a possibly-incorrect default.

The search API returns the v2 values; this doesn't break the existing search
frontend code since it already asks the server for the list of possible values.
But it does potentially cause search results to be different than before.