SW-1512 Add permissions for setting withdrawal user IDs
We're adding a "withdrawn by" field to withdrawals; managers, admins, and owners
can set it to the identity of any user in the organization, but contributors can
only set it to themselves.

Add two permissions to support this logic:

* `canReadOrganizationUser` is true if you have permission to read the membership
  information for a given user in a given organization.
* `canSetWithdrawalUser` is true if you have permission to set the user ID on a
  withdrawal for an accession.

No application code uses these new permissions yet; this is groundwork for a
subsequent change.