SW-931 Purge data when organization is deleted (#623)
When the last user is removed from an organization, delete its data.

Currently, a connected device manager counts as a user; if the organization has a
facility with a connected device manager, its data will not be removed from the
system when the last human user is removed.

Deployment note: The second database migration may take a while to run on the
production database, which may cause the deploy to look like it has gotten stuck.