Drop per-accession species details
In the initial version of the seed bank web app, we didn't have a concept of a
"species" per se; we just had some accession fields to hold information about the
seeds' species.

We have since moved over to a centrally-managed species list. The per-accession
species details are no longer referenced anywhere, so we can drop them from the
database and the API.