SW-1577 Add cut test support to v2 API (#630)
Change the internal representation of cut test results to the v2 style, where they
are represented as another kind of viability test. This new representation is used
even when accessions are written using the v1 API, but the v1 API continues to
present cut test results as a set of accession-level fields.

Existing cut test results are converted to viability tests by a database migration.

Clients can switch back and forth between the v1 and v2 APIs, but if an accession
has multiple cut tests (which is allowed in v2) they will be collapsed to a single
aggregate test when the accession is written using the v1 API.

The interaction between cut test results and accession quantity is a little
different than before: previously, we would only subtract seeds from the accession
if all three fields (compromised, empty, filled) were filled in with values. In
the new model, a withdrawal is created even if some of those values are still
null.

Deployment note: This change will require brief downtime. If a v1 PUT request with
cut test results is handled by an older server instance after the migration here is
finished, the edit will effectively be lost because the older code won't create a
new viability test.