Make "seeds tested" required in v2 viability tests (#634)
In v2, we don't allow users to create viability tests without saying how many
seeds they're testing. Enforce that requirement in the API.

The generated OpenAPI schema was not including the list of required fields in
the viability test payloads; this seems to be due to a bug in the OpenAPI library
that causes it to ignore the `required` attribute on write-only fields. Remove
the annotation that marked the POST and PUT payloads as write-only; it's more
important to have fields correctly marked as required.