SW-1721 Remove "Cleaning" accession state
Hazards of implementing a spec that's still in flux: originally the v2 seeds PRD
had a "Cleaning" state, and the code implements it. But that state was removed
from the PRD and replaced with the v1 "Processing" state.

Remove "Cleaning" from the list of valid states and make "Processing" valid in v2.