SW-1732 Preserve staff responsible in v2 API
In v1, viability tests and withdrawals have a "Staff Responsible" field which is
an arbitrary text value. In v2, we've replaced that field with a Terraware user
ID. But we don't want to lose any text the users entered into the field in v1,
so add it to the "Notes" field.

If the "Notes" field already contains the text of the "Staff Responsible" field,
we don't modify it; this prevents us from adding it multiple times if an accession
repeatedly flips between v1 and v2 writes.