Move seedbank tables to separate DB schema (#655)
To get ahead of name collisions when we add nursery support to the code, update
the jOOQ configuration to support generating code from multiple database schemas,
and move the seed-related tables to a new `seedbank` schema. Other tables remain
in the default schema (`public`).

This adds a new level of package hierarchy to the generated jOOQ classes.
Seed-related tables are now under a `seedbank` package. The tables in the
`public` schema are under `default_schema` because the word `public` is a
keyword that isn't a valid package name. The code generation config now breaks
the lists of tables and enum classes out by schema name.

For the most part, the impact on application code and tests is limited to
updating all the `import` statements that import jOOQ classes and fields. A small
amount of test code that uses JDBC to query table metadata needed to be changed
to be explicit about which schema to use.