SW-1365 Rename founderId to plantId in v2 accession API
In v1, we used the term "founder ID" to refer to an externally-issued identifier
for an individual specimen of an endangered plant species. We're switching to the
term "plant ID" in the v2 user interface. Update the API to match the new
terminology, and add the field to the search API using its new name.