SW-1803 Parse updated accession CSV template
The CSV importer was based on an early draft of the CSV template where the
instructions were on a separate row from the column names, but the final template
includes the instructions and column names in a single row, followed by a blank
row. Users might fill in values in the blank row, or they might not; handle either
situation by only skipping the first (header) row and ignoring blank rows
elsewhere in the file.