SW-1803 Map accession CSV country names to codes (#662)
In the template CSV we don't specify that the user has to feed us valid ISO
country codes in upper case, and requiring users to do that would be needlessly
annoying. Update the CSV importer to accept country names and lower-case country
codes, and add a validation check to detect unrecognized countries.