SW-1895 Add endpoint to create seedling batches
This change introduces the `nursery` database schema and creates an initial
version of the tables we'll need for our nursery features. It adds a new
API endpoint `POST /api/v1/nursery/batches` to create a new seedling batch,
and creates a bit of code infrastructure for subsequent nursery features.

A new test script can be used to create batches in bulk for manual testing.