SW-1845 Refactor accession number generation
We're going to need to allocate user-facing identifiers from the same pool of
values for accessions and seedling batches, as well as change them so they're
scoped per organization rather than globally.

This first step is to decouple the generation of these identifiers from the
seedbank code.

No change in functionality here, just moving the logic out of `AccessionStore`.