SW-1970 Add endpoint to delete seedling batches
`DELETE /api/v1/nursery/batches/{id}` will delete a batch from the system.

The ability to delete a batch means we have to relax some of our constraints.
For example, it becomes possible for there to be a nursery transfer withdrawal
where the destination batch no longer exists; we still want the withdrawal to
exist on the originating side.